requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.12.0
anthropic>=0.8.0
finnhub-python>=2.4.18
ta>=0.10.2  # Technical analysis
nltk>=3.8.1  # For sentiment analysis
scikit-learn>=1.3.0
plotly>=5.18.0  # For visualization
streamlit>=1.31.0  # For UI 